changed default types to my user-defined types

+++ include/renderer.h
@@ -4,10 +4,11 @@
 #include "player.h"
 #include "window.h"
 #include "map.h"
+#include "types.h"
 
 void draw_player(SDL_Renderer* renderer, player_t* player, map_t* map, SDL_Rect* map_rect);
 SDL_Texture* cache_map(SDL_Renderer* renderer, map_t* map);
 void draw_bg(SDL_Renderer* renderer);
-void draw_3d(SDL_Renderer* renderer, int colNum, float rayDist, int pixels, float rayA, player_t* player, map_t* map);
+void draw_3d(SDL_Renderer* renderer, usize colNum, f32 rayDist, u8 pixels, f32 rayA, player_t* player, map_t* map);
 
 #endif
